services/analytics-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
kafka-python==2.0.2
pymongo==4.6.0
pydantic==2.5.0
pydantic-settings==2.1.0
